Manually remove off-by-default RUF014 from schema
**Summary** Previously, `RUF014` would be part of ruff.schema.json depending on whether or not the `unreachable-code` feature was active. This caused problems for contributors who got unrelated RUF014 changes when updating the schema without the feature active.

This code in this PR is ugly (we're editing an automatically derived schema manually) and i'm happy about better suggestions. I thought about removing all feature gates except the one that actually runs the rule in the checker, but this would mean users could select the rule which would then silently not work. I also considered adding RUF014 always but the rule is not active in release builds so it should not be part of the schema imho.

**Test plan** `cargo dev generate-all` and `cargo run --bin ruff_dev --features unreachable-code -- generate-all` now have the same effect.